refactor: MenuAggregate 원복 및 Order와 OrderTable 재분리
